Added postgresql driver.
It should works mostly with any version of postgresql, but because of
the absence of lastval() function, you should use postgresql 8.1 or
better.

Because a couple of difference in sql dialect used in postgresql I
introduced some code to handle it in case of use it.
To check this, I introduced the driver name in rapyd class.

Should work correctly and leave the other driver code untouched.